Controllerlar Baştan Aşağı Güncellendi
Strateji bilgileri getir doğru çalışıyor. Veriler parça parça mı gelsin yoksa kırılım olarak mı gelsin belirlenmeli. ona göre object yapısı oluşturulmalı. Delete hala ilişkili datalar için doğru çalışmıyor.

+++ AKYSTRATEJI/Model/StIsler.cs
@@ -12,7 +12,7 @@ namespace AKYSTRATEJI.Model
         public DateTime? BaslangicTarihi { get; set; }
         public DateTime? BitisTarihi { get; set; }
         public DateTime OlusturmaTarihi { get; set; }
-        public short Deger { get; set; }
+        public int Deger { get; set; }
         public int? Ilce { get; set; }
         public int? Mahalle { get; set; }
         public int IslerId { get; set; }
