+++ AKYSTRATEJI/ViewModals/VMIsler.cs
@@ -13,7 +13,7 @@ namespace AKYSTRATEJI.ViewModals
         public DateTime? BaslangicTarihi { get; set; }
         public DateTime? BitisTarihi { get; set; }
         public DateTime OlusturmaTarihi { get; set; }
-        public short Deger { get; set; }
+        public int Deger { get; set; }
         public enums.Ilceler Ilce { get; set; }
         public enums.Mahalleler Mahalle { get; set; }
     }
